aviationstack-task/requirements.txt
requests
psycopg2-binary
dash
pandas
plotly